refactor: simplify TasksList and TaskItem components by removing unused code and improving task rendering; adjust layout in Tasks page
